backend_and_parser/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.11
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.10
lxml==4.6.1
mysqlclient==2.0.1
parse==1.18.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
sentry-sdk==0.19.2
django-cors-headers==3.5.0